Added a very simple example, and re-titled to match the goal, not the means.
The developer's perspective is "I know what page I want the user to see.
How do I make that happen with couchdb?".  The answer is to organize the
data in a view function, and then write a list function which will do the content-type
specific parts, such as generate HTML.  This page presumed the user already
knew that they needed to transform a view to get content, and was
already thinking from that perspective.  I hope my addition at the top
brings them across, before delving into details.