Disable arrow-based navigation when modifiers are used
On Mac Firefox (at least) the keyboard shortcut to switch next/previous tab is Cmd-Alt-{Left,Right}. The navigation script in the guide only sees that there's a Left/Right key being pressed and navigates on every tab switch. This leads to totally maddening behaviour whereby switching between the guide and another tab next to it (which is typical in development) causes the guide to navigate. Checking for modifiers fixes this.